refactor(be): move VerifiedEmail to its own type module
The interface struct lived alongside EmailRecoveryCredential in
`types::email_recovery` even though the two anchor primitives are
parallel — same leftover coupling the canister-side hoist into
`email_inbound` already fixed for the Rust modules. Now matches.

New module `types::verified_email` owns `VerifiedEmail`; re-exported
through the workspace's existing top-level glob so any caller using
the flat `internet_identity_interface::internet_identity::types::*`
import keeps working. Explicit-path importers updated to the new
location.
